IOException when reading/sending file > 2GB. now using uint instead of int in file-seek.

+++ FO-DICOM.Core/IO/StreamByteSource.cs
@@ -139,12 +139,12 @@ namespace FellowOakDicom.IO
             else if (count >= LargeObjectSize && _readOption == FileReadOption.ReadLargeOnDemand)
             {
                 buffer = new StreamByteBuffer(_stream, _stream.Position, count);
-                _stream.Seek((int)count, SeekOrigin.Current);
+                _stream.Seek(count, SeekOrigin.Current);
             }
             else if (count >= LargeObjectSize && _readOption == FileReadOption.SkipLargeTags)
             {
                 buffer = null;
-                Skip((int)count);
+                Skip(count);
             }
             else // count < LargeOpjectSize || _readOption == FileReadOption.ReadAll
             {
@@ -157,7 +157,7 @@ namespace FellowOakDicom.IO
         public Task<IByteBuffer> GetBufferAsync(uint count) => Task.FromResult(GetBuffer(count));
 
         /// <inheritdoc />
-        public void Skip(int count) => _stream.Seek(count, SeekOrigin.Current);
+        public void Skip(uint count) => _stream.Seek(count, SeekOrigin.Current);
 
         /// <inheritdoc />
         public void Mark() => _mark = _stream.Position;
